Reorganized the library to comply with arduino 1.5+ library standards https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification
