+++ library.properties
@@ -1,5 +1,5 @@
 name=firebase-arduino
-version=0.4
+version=0.5
 author=
 maintainer=
 sentence=Fill in later
