+++ src/FirebaseHttpClient.h
