+++ src/FirebaseHttpClient_Esp8266.cpp
