+++ src/modem/diagram.png
