+++ src/modem/input-stream.h
