+++ src/modem/output-stream.h
