+++ test/modem/mock-input-stream.h
