+++ test/modem/stream-command_test.cpp
@@ -1,8 +1,8 @@
 #include "Firebase.h"
 #include "gtest/gtest.h"
 #include "modem/commands.h"
-#include "modem/test/mock-input-stream.h"
-#include "modem/test/mock-output-stream.h"
+#include "test/modem/mock-input-stream.h"
+#include "test/modem/mock-output-stream.h"
 #include "test/mock-firebase.h"
 
 namespace firebase {
